Add shink ('@') and expand ('*') column sizing
It a <columns>'s `template` attribute, now three specifier types are supported:

1) `\\d+%?`: aka a pos. int., w/ optional ending '%'; as before, denotes a % of the document width.

2) `@`: column will take the absolute size of it's largest child.
	- <item>, <image>, and <entity> all have natural widths
  - <text> (and hence <h1>, <h2>, and <p>) now take a `minWidth` attribute to work with such a column spec.
	- The `TagElement` interface now has a getWidth() function; by default implementation it returns 0, like getHeight() does.

3) `*`: column will share with any other `*` columns an equal portion of any remaining un-specified document width.

In the current code structure, render() is the only time a <column> has access to it's children's widths.
There is a boolean check to have this setup run only once.

<columns> itself also has an override for getWidth(), as experience has shown nested columns to be useful.

Incidentally, the `elements` instance variable has been renamed to `children` to be consistent with other tags.